requirements.txt
absl-py==1.4.0
asttokens==2.4.0
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
comm==0.1.4
debugpy==1.7.0
decorator==5.1.1
deephaven-core==0.28.0
deephaven-ipywidgets==0.3.2
deephaven-plugin==0.5.0
deephaven-server==0.28.0
executing==1.2.0
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.9.0
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
ipywidgets==8.1.0
java-utilities==0.2.0
jedi==0.18.2
jpy==0.14.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
keras==2.13.1
libclang==16.0.6
llvmlite==0.40.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numba==0.57.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stack-data==0.6.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.7
widgetsnbextension==4.0.8
wrapt==1.15.0
